# Various script improvements for git clone calls